requirements.txt
streamlit>=1.32.0
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=10.0.0
plotly>=5.18.0
gdown>=5.1.0
